requirements.txt
joblib==1.3.2
numpy==1.26.4
pandas==1.5.3
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
spark==0.2.1
threadpoolctl==3.3.0
